% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{activate_conda_env}
\alias{activate_conda_env}
\title{Activate conda environment}
\usage{
activate_conda_env(name = NULL, prefix = NULL, conda_home = "~/miniconda3")
}
\arguments{
\item{name}{Name of the conda environment}

\item{prefix}{Path to the environment}

\item{conda_home}{Path to the miniconda installation}
}
\description{
Add conda environment to PATH to call programs not available via R
}
