% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annotate_PhANNs}
\alias{annotate_PhANNs}
\title{Prokka}
\usage{
annotate_PhANNs(
  genome.fasta = NULL,
  output.dir = NULL,
  conda.env = "../envs/phanns",
  repo.dir = "../PhANNs",
  repo.url = "https://github.com/Adrian-Cantu/PhANNs.git",
  model.url = "https://edwards.sdsu.edu/phanns/download/model.tar"
)
}
\arguments{
\item{genome.fasta}{Path to genome FASTA}
}
\description{
...
PhANNs
}
\details{
Annotate phage structural proteins
}
