% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Annotation-class}
\alias{Annotation-class}
\title{The Annotation class}
\description{
The Annotation object stores file paths for genome annotation The core components are
the genes GFF file, the coding sequences (CDS) FASTA, the protein FASTA.
}
\section{Slots}{

\describe{
\item{\code{index}}{A character vector sample indices matching index(genomeCollection)}

\item{\code{genes}}{A character vector with paths to the genes GFF.}

\item{\code{cds}}{A character vector with paths to the coding sequences FASTA.}

\item{\code{proteins}}{A character vector with paths to the protein FASTA.}

\item{\code{log}}{A character vector with paths to the annotation log file.}

\item{\code{genome}}{A character vector indicating the column in metadata(genomeCollection) storing genome locations.}

\item{\code{tool}}{A character vector indicating the annotation tool (e.g. Bakta, Padloc, DefenseFinder, ...)}

\item{\code{type}}{A character vector indicating the annotation type}
}}

