% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_genomeCollection.R
\name{Annotation,genomeCollection-method}
\alias{Annotation,genomeCollection-method}
\title{Accessors to Assembly objects contained in a genomeCollection object.}
\usage{
\S4method{Annotation}{genomeCollection}(x, name = NULL)
}
\description{
Accessors to Assembly objects contained in a genomeCollection object.
}
