% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_genomeCollection.R
\name{Assembly,genomeCollection-method}
\alias{Assembly,genomeCollection-method}
\title{Accessors to Assembly objects contained in a genomeCollection object.}
\usage{
\S4method{Assembly}{genomeCollection}(x, name = NULL)
}
\description{
Accessors to Assembly objects contained in a genomeCollection object.
}
