% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{assign_phage_taxonomy_by_ICTV}
\alias{assign_phage_taxonomy_by_ICTV}
\title{Infer taxonomic ranks for phage genomes from International Committee for Taxonomy of Viruses (ICTV) database}
\usage{
assign_phage_taxonomy_by_ICTV()
}
\description{
Infer taxonomic ranks for phage genomes from International Committee for Taxonomy of Viruses (ICTV) database
}
