% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{bakta}
\alias{bakta}
\title{Bakta annotation}
\usage{
bakta(
  input.fasta = NULL,
  output.dir = NULL,
  output.prefix = NULL,
  db = NULL,
  download.db = FALSE,
  threads = n_proc(),
  force = FALSE
)
}
\arguments{
\item{input.fasta}{Character, path to FASTA file}

\item{output.dir}{Character, path to output dir}

\item{output.prefix}{Character,}

\item{db}{Character, path to the bakta database}

\item{download.db}{Boolean, whether to download the database if missing}

\item{threads}{Integer, number of threads to use (default: all)}

\item{force}{Boolean, whether to force overwriting existing output folder}
}
\description{
CLI wrapper for genome annotation using Bakta
}
