% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_gaudi.R
\name{c_gaudi}
\alias{c_gaudi}
\title{Consensus Group Aggregation via UMAP Data Integration}
\usage{
c_gaudi(
  omics,
  n_max = 10,
  umap_params = list(n_neighbors = 15, n_components = 4),
  umap_params_conc = list(n_neighbors = 15, n_components = 2),
  min_pts = NULL,
  xgboost_params = list(lambda = 0, eta = 1, gamma = 100, max_depth = 10, subsample =
    0.95),
  compute_features = TRUE,
  combine_omics = FALSE,
  clean_feature_names = FALSE,
  samples_in_rows = TRUE,
  reassign_cluster_zero = FALSE,
  method = "xgboost"
)
}
\arguments{
\item{omics}{A \code{list} of omics data frames or matrices.}

\item{n_max}{Numeric value indicating the total number of latent spaces to be computed (the final latent space will be based on all these spaces).}

\item{umap_params}{Parameters for UMAP factorization of individual omics datasets.}

\item{umap_params_conc}{Parameters for UMAP factorization of concatenated omics data.}

\item{min_pts}{Minimum number of points per cluster in HDBSCAN clustering.}

\item{xgboost_params}{Parameters for XGBoost model used in metagenes computation.}

\item{compute_features}{Logical indicating whether to compute metagenes.}

\item{combine_omics}{Logical indicating whether to combine combined omics metagenes.}

\item{clean_feature_names}{Logical indicating whether to clean feature names in omics datasets.}

\item{samples_in_rows}{Logical indicating whether samples are in rows (default) or columns.}

\item{reassign_cluster_zero}{Logical indicating whether to reassign data points in cluster 0.}

\item{method}{Character indicating the regression model used to extract features. Options are "xgboost" and "rf".}
}
\value{
An object of class \code{GAUDIObject} containing the results of the multi-omics integration and analysis.
This includes factors, clusters, silhouette scores, individual factors, metagenes, and the \code{gaudi} package version.
}
\description{
\code{c_gaudi} performs unsupervised integration of multi-omics data.
It utilizes Uniform Manifold Approximation and Projection (UMAP) for dimensionality reduction and HDBSCAN for clustering.
The method concatenates individual UMAP embeddings of each omics data type, followed by a second UMAP for integrated analysis.
Clustering is performed on the integrated data, and optionally, cluster 0 reassignment and feature computation (metagenes) can be done.
}
\examples{
omics_list <- list(data.frame(matrix(rnorm(200), ncol = 20)),
                   data.frame(matrix(rnorm(200), ncol = 20)))
c_gaudi_result <- c_gaudi(omics_list, n_max = 3, samples_in_rows = FALSE)

}
