% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_eop}
\alias{calc_eop}
\title{Calculate efficiency of plating/plaquing (EOP) from plaque forming units (PFU)}
\usage{
calc_eop(test.pfu = NULL, iso.pfu = NULL)
}
\arguments{
\item{test.pfu}{Numeric, vector of PFU units}

\item{iso.pfu}{Numeric, vector of PFU units}
}
\description{
Calculate efficiency of plating/plaquing (EOP) from plaque forming units (PFU)
}
