% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_pfu}
\alias{calc_pfu}
\title{Calculate plaque forming units (PFU) from spot assay}
\usage{
calc_pfu(dil.factor = NULL, plaque.count = 10, droplet.volume = 0.0025)
}
\arguments{
\item{dil.factor}{Dilution factor of the droplet}

\item{plaque.count}{Number of plaques counted from the droplet}

\item{droplet.volume}{Volume of the deposited droplet (mL)}
}
\description{
Calculate PFU from plaque counts observed for a dilution of phage stock. Requires the spot volume.
}
