% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_Annotation.R
\name{$,Annotation-method}
\alias{$,Annotation-method}
\title{Data}
\usage{
\S4method{$}{Annotation}(x, name)
}
\arguments{
\item{x}{An Annotation object}

\item{name}{Name of Annotation slot}
}
\description{
Get Annotation data
}
