% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_Assembly.R
\name{$,Assembly-method}
\alias{$,Assembly-method}
\title{Data}
\usage{
\S4method{$}{Assembly}(x, name)
}
\arguments{
\item{x}{A Assembly object}

\item{name}{Name of Assembly slot}
}
\description{
Get Assembly data
}
