% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_genomeCollection.R
\name{$,genomeCollection-method}
\alias{$,genomeCollection-method}
\title{Metadata}
\usage{
\S4method{$}{genomeCollection}(x, name)
}
\arguments{
\item{x}{A genomeCollection object}

\item{name}{Name of genome meta data}
}
\description{
Get and set genome meta data
}
