% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_installed}
\alias{check_installed}
\title{Check if program is installed on the system}
\usage{
check_installed(program = NULL, silent = FALSE)
}
\arguments{
\item{program}{Name of the program}

\item{silent}{Whether to report what is found}
}
\description{
Helper function to determine if programs are installed (available) on the system.
Missing programs lead to an error that are intended to break higher-level functions.
}
