% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_version}
\alias{check_version}
\title{Check program version}
\usage{
check_version(program = NULL, command = "--version")
}
\arguments{
\item{program}{Name of the program}
}
\description{
Check program version
}
