% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colorscale_diverging}
\alias{colorscale_diverging}
\title{Create diverging colorscale}
\usage{
colorscale_diverging(
  x = NULL,
  midpoint = 0,
  breaks_per_side = 50,
  palette = "RdYlBu",
  direction = -1
)
}
\arguments{
\item{x}{Numeric vector or matrix}

\item{midpoint}{Integer within data range}

\item{breaks_per_side}{Integer specifying the number of color values on either side of the midpoint}

\item{palette}{Name of the RColorBrewer palette}

\item{direction}{Direction of the color values (1, -1)}
}
\description{
Creates a diverging colorscale around a midpoint
based on RColorBrewer palettes
}
