% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{CombineAnnotations}
\alias{CombineAnnotations}
\title{Combine annotations for a genomeCollection}
\usage{
CombineAnnotations(
  object,
  name,
  annotations,
  db.amrfinder = NULL,
  db.phrogs = NULL,
  remove.duplicates = FALSE,
  remove.source = NULL,
  overwrite = FALSE,
  debug.sample = NULL
)
}
\arguments{
\item{object}{A genomeCollection object}

\item{name}{Name of the slot in metadata(object)}

\item{annotations}{A character vector of annotations present in Annotations(object)}

\item{db.amrfinder}{Path to AMRfinder database, linking HMM model names to AMR types (see https://www.ncbi.nlm.nih.gov/pathogens/hmm/#).}

\item{remove.duplicates}{Whether to remove duplicated entries from the final GFF file. Order of the input GFFs will be preserved.
See help(format_gff3) for more information.}

\item{remove.source}{I came back to this after some time and have no idea what I intended (and no description). UPDATE !!!}

\item{overwrite}{Whether to overwrite existing output files}
}
\description{
Combine the genes from multiple Annotation objects in a genomeCollection.
}
