% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_quality.R
\name{CountFastq}
\alias{CountFastq}
\title{CountFastq}
\usage{
CountFastq(object, read.col = "R1", threads = n_proc())
}
\arguments{
\item{object}{genomeCollection}

\item{read.col}{Column in metadata(object) containing FASTQ paths}

\item{threads}{Integer, number of threads to use}
}
\description{
Count FASTQ files for a genomeCollection
}
