% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_quality.R
\name{CountFastqPlot}
\alias{CountFastqPlot}
\title{CountFastqPlot}
\usage{
CountFastqPlot(object, read.col = "R1")
}
\arguments{
\item{object}{genomeCollection}

\item{read.col}{Column in metadata(object) containing FASTQ paths}
}
\description{
View FASTQ counts
}
