% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_Reads.R
\name{Counts,Reads-method}
\alias{Counts,Reads-method}
\title{Accessors for the FASTQ quality summaries of a Reads object.}
\usage{
\S4method{Counts}{Reads}(x)
}
\description{
Accessors for the FASTQ quality summaries of a Reads object.
}
