% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_genomeCollection.R
\name{genomeCollection}
\alias{genomeCollection}
\title{Create a \code{genomeCollection}}
\usage{
genomeCollection(
  ids,
  path,
  meta.data = NULL,
  path.key = "path",
  genome.key = "genome",
  genes.key = "genes",
  cds.key = "CDS",
  proteins.key = "proteins"
)
}
\arguments{
\item{path}{Character of length 1, path to a}

\item{meta.data}{A data.frame containing meta data associated to ids}
}
\value{
A \code{\link{genomeCollection}} object
}
\description{
Create a \code{genomeCollection} object from a data.frame with defined components.
}
