% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Distance-class}
\alias{Distance-class}
\title{The Distance class}
\description{
The Distance object stores file paths for genome distances to infer phylogeny.
The core components are ...
}
\section{Slots}{

\describe{
\item{\code{index}}{A character vector sample indices matching index(genomeCollection)}

\item{\code{distance}}{...}

\item{\code{sketches}}{...}

\item{\code{log}}{A character vector with paths to the annotation log file.}

\item{\code{tool}}{A character vector indicating the distance tool (e.g. Mash, FastANI, skani, MMseqs, MUMer, Foldseek, ...)}

\item{\code{type}}{A character vector indicating the distance type (e.g. genome, CDS, protein, ...)}
}}

