% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_Annotation.R
\name{.DollarNames.Annotation}
\alias{.DollarNames.Annotation}
\title{Dollar-sign autocompletion}
\usage{
\method{.DollarNames}{Annotation}(x, pattern = "")
}
\description{
Dollar-sign autocompletion
}
