% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_genomeCollection.R
\name{.DollarNames.genomeCollection}
\alias{.DollarNames.genomeCollection}
\title{Dollar-sign autocompletion}
\usage{
\method{.DollarNames}{genomeCollection}(x, pattern = "")
}
\description{
Dollar-sign autocompletion
}
