% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_clusters}
\alias{drop_clusters}
\title{Drop Specified Clusters}
\usage{
drop_clusters(object, clusters = 0)
}
\arguments{
\item{object}{A \code{GAUDIObject}.}

\item{clusters}{A vector of cluster IDs (numeric or character) that should be removed from the object.
Default is 0, which means it will remove the cluster(s) labeled as '0'.}
}
\value{
An object similar to the input but with the specified clusters and their corresponding data removed.
The function performs in-place modification but also returns the modified object for convenience.
If the object is not valid after modification, the function will not return it.
}
\description{
This function removes specified clusters from an \code{GAUDIObject}.
}
