% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{entrez_search_count_by_year}
\alias{entrez_search_count_by_year}
\title{Find NCBI database entries for a topic per year}
\usage{
entrez_search_count_by_year(db = "nuccore", year = NULL, term = NULL)
}
\arguments{
\item{db}{NCBI database to query}

\item{year}{Year to search entries for}

\item{term}{Search term}
}
\description{
Find NCBI database entries for a topic per year
}
