% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{extract_gff3_attributes}
\alias{extract_gff3_attributes}
\title{Extract attributes from GFF3}
\usage{
extract_gff3_attributes(object = NULL, attribute_key = "attribute_")
}
\arguments{
\item{object}{Data.frame formatted as GFF3}

\item{attribute_key}{String appended to columns to indicate they are extracted from 'attributes'}
}
\value{
data.frame
}
\description{
Extract values from GFF3 attributes and add columns to object
}
