% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{extract_tag_value_pairs}
\alias{extract_tag_value_pairs}
\title{Extract tag-value pairs by pattern}
\usage{
extract_tag_value_pairs(x, pattern = "=", verbose = FALSE)
}
\arguments{
\item{x}{A character string containing tag-value pairs, sepearated by pattern}

\item{pattern}{A string separating tags from values (e.g. ID=myname)}
}
\description{
Extract tag-value pairs by pattern
}
