% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_quality.R
\name{fastp}
\alias{fastp}
\title{FastP filtering of short reads}
\usage{
fastp(
  input_read_1 = NULL,
  output_read_1 = NULL,
  input_read_2 = NULL,
  output_read_2 = NULL,
  qualified_quality_phred = 15,
  unqualified_percent_limit = 40,
  length_required = 15,
  average_quality = 0,
  threads = n_proc(),
  overwrite = FALSE
)
}
\arguments{
\item{input_read_1}{Character, Input file name.}

\item{output_read_1}{Character, Output file name.}

\item{input_read_2}{Character, Input file name.}

\item{output_read_2}{Character, Output file name.}

\item{qualified_quality_phred}{Numeric, the quality value that a base is qualified.}

\item{unqualified_percent_limit}{Numeric, how many percents of bases are allowed to be unqualified (0~100).}

\item{length_required}{Numeric, shorter reads will be discarded.}

\item{average_quality}{Numeric, if one read's average quality score is lower, then this read/pair is discarded.}

\item{threads}{Integer, number of worker threads}

\item{overwrite}{Boolean, whether to overwrite output.}
}
\description{
CLI wrapper for FastP to filter and trim short reads
}
