% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads.R
\name{fastqc}
\alias{fastqc}
\title{FastQC}
\usage{
fastqc(files = NULL, out.dir = NULL, threads = n_proc())
}
\arguments{
\item{files}{Character vector of FASTQ files}

\item{out.dir}{Path to output directory}

\item{threads}{Number of cores to use}
}
\description{
CLI wrapper for FastQC
}
