% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads.R
\name{FilterReads}
\alias{FilterReads}
\title{FilterReads}
\usage{
FilterReads(
  object,
  name.from = "raw",
  name.to = "filtered",
  fastp.length_required = 40L,
  fastp.average_quality = 20L,
  fastp.unqualified_percent_limit = 40L,
  fastp.qualified_quality_phred = 15L,
  fastplong.length_required = 100L,
  fastplong.mean_qual = 20L,
  fastplong.unqualified_percent_limit = 40L,
  fastplong.qualified_quality_phred = 15L,
  recompute = FALSE,
  recompute.sample = NULL
)
}
\arguments{
\item{object}{genomeCollection}

\item{name.from}{Name of the Reads object storing unfiltered FASTQ files}

\item{name.to}{Name of the Reads object to create for filtered FASTQ files}

\item{fastp.length_required}{Numeric, see ?fastp()}

\item{fastp.average_quality}{Numeric, see ?fastp()}

\item{fastp.unqualified_percent_limit}{Numeric, see ?fastp()}

\item{fastp.qualified_quality_phred}{Numeric, see ?fastp()}

\item{fastplong.length_required}{Numeric, see ?fastplong()}

\item{fastplong.mean_qual}{Numeric, see ?fastplong()}

\item{fastplong.unqualified_percent_limit}{Numeric, see ?fastplong()}

\item{fastplong.qualified_quality_phred}{Numeric, see ?fastplong()}
}
\description{
Filter FASTQ reads using fastp/fastplong
}
