% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{format_gff3}
\alias{format_gff3}
\title{Enforce GFF3 formatting}
\usage{
format_gff3(x, replace.attributes = FALSE)
}
\arguments{
\item{x}{Data.frame}
}
\description{
Format a data.frame according to GFF3 structure.
}
