% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{gaudi_enrichment}
\alias{gaudi_enrichment}
\title{Enrichment Analysis on GAUDI Object}
\usage{
gaudi_enrichment(
  object,
  on_omics,
  on_factor = 1,
  organism = org.Hs.eg.db,
  ontology = "ALL",
  minGSSize = 2,
  maxGSSize = 200
)
}
\arguments{
\item{object}{A \code{GAUDIObject}.}

\item{on_omics}{Numeric index indicating which omics dataset in the \code{object} to use for plotting. Names of metagenes here should be \strong{gene symbols}.}

\item{on_factor}{Numeric index indicating which factor from the multi-omics integration to use for plotting.}

\item{organism}{Character string specifying the organism database. Default is "org.Hs.eg.db".}

\item{ontology}{One of "BP", "MF", and "CC" subontologies, or "ALL" for all three}

\item{minGSSize}{Minimum gene set size}

\item{maxGSSize}{Maximum gene set size}
}
\value{
A data frame containing the enrichment results.
}
\description{
This function performs enrichment analysis on GAUDI results using the
'clusterProfiler' package's Gene Set Enrichment Analysis (GSEA).
}
