% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{GAUDIObject-class}
\alias{GAUDIObject-class}
\title{GAUDIObject Class}
\description{
A class representing the output of the Group Aggregation via UMAP Data Integration (GAUDI) method.
The \code{GAUDIObject} class encapsulates various results and metadata from the GAUDI analysis,
including integrated factors, clustering results, silhouette scores, and metagene associations.
}
\details{
The \code{GAUDIObject} class is designed to store comprehensive results from the GAUDI analysis, providing an integrated view of multi-omics data.
It allows for easy access to both combined and individual omics data results, facilitating further analysis and visualization.
}
\section{Slots}{

\describe{
\item{\code{factors}}{A \code{data.frame} containing the UMAP factors for the integrated multi-omics data.}

\item{\code{clusters}}{A numeric vector representing the cluster assignments for each sample.}

\item{\code{silhouette_score}}{A numeric value representing the average silhouette score of the clustering.}

\item{\code{individual_factors}}{A list of UMAP factors for individual omics datasets.}

\item{\code{metagenes}}{A list containing metagenes for individual omics datasets.}

\item{\code{gaudiVersion}}{A character string indicating the version of the GAUDI package used for the analysis.}
}}

