% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_batch}
\alias{generate_batch}
\title{Generate and Execute Repeated Elements Vector}
\usage{
generate_batch(vec)
}
\arguments{
\item{vec}{A numeric vector where each element indicates the number of times
the corresponding index should be repeated in the output vector.}
}
\value{
A vector where each integer from 1 to the length of the input vector
is repeated according to the values in the input vector.
}
\description{
This function takes a numeric vector and generates a new vector
where each element is repeated according to the corresponding value in the input vector.
}
\examples{
# Example usage
vec <- c(3, 2, 5)
generate_batch(vec)

}
