% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_Annotation.R
\name{Genes,Annotation-method}
\alias{Genes,Annotation-method}
\title{Accessors for the components of an Annotation object.}
\usage{
\S4method{Genes}{Annotation}(x)
}
\description{
Accessors for the components of an Annotation object.
}
\author{
Oliver Dietrich
}
