% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{genomeCollection-class}
\alias{genomeCollection-class}
\title{The genomeCollection Class}
\description{
The genomeCollection object contains genome sequence data. Formatted based
on NCBI datasets.
}
\section{Slots}{

\describe{
\item{\code{index}}{Character, unique identifiers for each genome}

\item{\code{path}}{Character, path to the directory where progress is tracked}

\item{\code{meta.data}}{Data.frame containing genome metadata}

\item{\code{assembly}}{A list of assembly objects}

\item{\code{annotation}}{A list of annotation objects}

\item{\code{distances}}{A list of distance objects}

\item{\code{misc}}{A list of miscellaneous information}
}}

