% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_genomeCollection.R
\name{index,genomeCollection-method}
\alias{index,genomeCollection-method}
\title{Accessors for the 'index' element of an genomeCollection object.}
\usage{
\S4method{index}{genomeCollection}(x)
}
\description{
The \code{index} slot in an genomeCollection object holds
a character vector containing unique genome names.
}
\examples{
index(object)
}
\author{
Oliver Dietrich
}
