% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{interactionSet-class}
\alias{interactionSet-class}
\title{The interactionSet Class}
\description{
The interactionSet object contains an interaction matrix from two genomeCollections.
}
\section{Slots}{

\describe{
\item{\code{interactions}}{Matrix of interactions}

\item{\code{phages}}{genomeCollection of bacteriophages}

\item{\code{bacteria}}{genomeCollection of bacteria}
}}

