% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_file}
\alias{is_file}
\title{Check if object refers to a single file}
\usage{
is_file(file_name = NULL, suffix = NULL, silent = FALSE)
}
\arguments{
\item{file_name}{Character, file path}

\item{suffix}{Character, file extension (e.g. '.csv','.fasta')}

\item{silent}{Boolean, whether to print warnings}
}
\description{
Check if object refers to a single file
}
