% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_valid_fastq}
\alias{is_valid_fastq}
\title{Check FASTQ formatting}
\usage{
is_valid_fastq(
  file = NULL,
  suffix = c(".fastq", ".fastq.gz"),
  multiple = FALSE
)
}
\arguments{
\item{file}{File name}

\item{suffix}{List of allowed file suffixes}
}
\description{
Check FASTQ formatting
}
