% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{jaccard}
\alias{jaccard}
\title{Compute jaccard similarity/distance}
\usage{
jaccard(a = NULL, b = NULL, type = "similarity")
}
\arguments{
\item{a}{Character vector}

\item{b}{Character vector}

\item{type}{Character, what type should be reported (similarity, distance)}
}
\value{
Numeric
}
\description{
Compute jaccard similarity of two character vectors.
}
