% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_mer_utils.R
\name{jellyfish_kmer_count}
\alias{jellyfish_kmer_count}
\title{Jellyfish k-mer counting}
\usage{
jellyfish_kmer_count(
  input.sequence = NULL,
  out.dir = NULL,
  k = 32,
  hash.size = 64,
  threads = n_proc(),
  remove.intermediates = TRUE
)
}
\arguments{
\item{input.sequence}{File path}

\item{k}{Integer, length of k-mer}

\item{hash.size}{Integer, Initial hash size}

\item{threads}{Integer, Number of threads}

\item{output.tsv}{File path}
}
\description{
CLI wrapper for Jellyfish to count k-mers
}
