% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{map_reads_to_reference}
\alias{map_reads_to_reference}
\title{Minimap2}
\usage{
map_reads_to_reference(
  reference,
  out.dir,
  short.1 = NULL,
  short.2 = NULL,
  unpaired = NULL,
  long = NULL,
  tool = "minimap2",
  preset = "map-ont",
  file.prefix = NULL,
  remove.intermediates = TRUE,
  recompute = FALSE,
  threads = n_proc()
)
}
\arguments{
\item{reference}{FASTA file}

\item{out.dir}{Path to output directory}

\item{short.1}{FASTQ file of first short reads in each pair (R1)}

\item{short.2}{FASTQ file of second short reads in each pair (R2)}

\item{unpaired}{FASTQ file of unpaired short reads (S)}

\item{long}{FASTQ or FASTA file of long reads}

\item{tool}{Name of alignment tool (minimap2 or bowtie2)}

\item{preset}{Preset for tuning the minimap 2 params. One of 'map-ont', 'map-pb', or 'map-ilcr' for long reads.
Will default to 'sr' for short reads.}

\item{threads}{Number of threads}
}
\description{
Minimap2 alignment of FASTQ reads to FASTA reference.
}
