% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{mash_screen}
\alias{mash_screen}
\title{Mash screen}
\usage{
mash_screen(
  query,
  reference,
  out.dir,
  k = 21,
  s = 1000,
  threads = n_proc(),
  recompute = FALSE
)
}
\description{
Screen the containment of a query in a set of reference sequences.
}
