% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{mash_sketch}
\alias{mash_sketch}
\title{Create mash sketch from sequences}
\usage{
mash_sketch(
  input,
  out.dir,
  prefix,
  k = 21,
  s = 1000,
  alphabet = "nucleotide",
  threads = n_proc(),
  recompute.sketch = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{input}{XStringSet or vector of FASTA file paths}

\item{prefix}{Name added as file prefix (e.g. query, reference)}

\item{k}{Integer, k-mer size}

\item{s}{Integer, sketch size}

\item{output.dir}{Path to output directory}
}
\description{
Create mash sketch from sequences
}
