% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{mash_sketch_distance}
\alias{mash_sketch_distance}
\title{Compute Mash distance between pre-computed sketches}
\usage{
mash_sketch_distance(
  ref.sketch = NULL,
  query.sketch = NULL,
  out.file = NULL,
  threads = n_proc()
)
}
\arguments{
\item{ref.sketch}{Character, path to reference sketch}

\item{query.sketch}{Character, path to query sketch}

\item{out.file}{Character, file name to output.tsv}
}
\description{
Compute Mash distance between pre-computed sketches
}
