% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly.R
\name{medaka_polish}
\alias{medaka_polish}
\title{Medaka polishing}
\usage{
medaka_polish(
  input.fastx = NULL,
  input.assembly = NULL,
  out.dir = NULL,
  fill_gaps = TRUE,
  model = "r1041_e82_400bps_sup_v5.0.0",
  threads = 1
)
}
\arguments{
\item{input.fastx}{FASTx input basecalls}

\item{input.assembly}{FASTA input assembly}

\item{out.dir}{Output folder}

\item{fill_gaps}{Boolean, whether to fill gaps in consensus with draft sequence}

\item{model}{Medaka model (see "medaka_consensus -h" for choices)}

\item{threads}{Number of threads with which to create features}

\item{batch_size}{Numeric, select batch size to control memory use}
}
\description{
CLI wrapper for Medaka polishing of genome assemblies.
Medaka is optimized to work with the Flye assembler.
}
