% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_mer_counts.R
\name{mercat_count_kmers}
\alias{mercat_count_kmers}
\title{Count k-mers using mercat2}
\usage{
mercat_count_kmers(sequence = NULL)
}
\arguments{
\item{sequence}{}
}
\description{
Count k-mers using mercat2
}
