% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_genomeCollection.R
\name{metadata,genomeCollection-method}
\alias{metadata,genomeCollection-method}
\title{Accessors for the 'meta.data' element of an genomeCollection object.}
\usage{
\S4method{metadata}{genomeCollection}(x)
}
\description{
The \code{meta.data} slot in an genomeCollection object holds
a data.frame containing unstructured information associated to each genome.
}
\examples{
metadata(object)
}
\author{
Oliver Dietrich
}
