% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{mummer_alignment}
\alias{mummer_alignment}
\title{MUMmer alignment}
\usage{
mummer_alignment(
  query,
  reference,
  out.dir,
  program = "nucmer",
  breaklen = if (program == "nucmer") 200 else 60,
  mincluster = if (program == "nucmer") 65 else 20,
  minmatch = if (program == "nucmer") 20 else 6,
  maxgap = if (program == "nucmer") 90 else 30,
  prefix = program,
  recompute = FALSE
)
}
\arguments{
\item{query}{FASTA file}

\item{reference}{FASTA file}

\item{out.dir}{Path to output directory}

\item{program}{Which mummer program to use (nucmer or promer)}

\item{breaklen}{Distance an alignment extension will attempt to extend poor scoring regions before giving up}

\item{mincluster}{Minimum length of a cluster of matches}

\item{minmatch}{Minimum length of a single match}

\item{maxgap}{Maximum gap between two adjacent matches in a cluster}

\item{prefix}{Character prefix to temporary files}

\item{recompute}{Wether to delete the out.dir and recompute}

\item{threads}{Number of threads}
}
\description{
MUMmer alignment
}
