% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{n_proc}
\alias{n_proc}
\title{Check output file}
\usage{
n_proc()
}
\description{
Helper function to determine if an output file exists.
Check output directory
}
\details{
Helpter function to determine if an output directory exists or should be created
Detect the number of processing units available
}
