% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{navigate_to_project_root}
\alias{navigate_to_project_root}
\title{Navigate to project root}
\usage{
navigate_to_project_root(
  project.name = NULL,
  subdirs = c("/bin", "/analysis", "/data", "/scripts")
)
}
\arguments{
\item{project.name}{Name of the project, needs to be part of the PATH}

\item{subdirs}{Character, Name of common subdirectories that can be recognized without project.name}
}
\description{
Navigate to the root of a project directory. The directory itself can be anywhere in a filesystem
but if e.g. Jupyter starts from a subdirectory this function navigates back to enable setting relative PATHs.
}
