% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_datasets.R
\name{ncbi_datasets_download_genome}
\alias{ncbi_datasets_download_genome}
\title{Download genome sequence data from NCBI using Assembly.version IDs}
\usage{
ncbi_datasets_download_genome(
  Assembly.version = NULL,
  out.dir = NULL,
  debug = FALSE,
  debug.n = 10
)
}
\arguments{
\item{accession}{Character vector with Assembly.version numbers}

\item{inputfile}{File with accession numbers (one per line)}
}
\description{
Wrapper for the NCBI datasets CLI
Based on the call 'datasets download genome accession' it retrieves sequence data
for Assembly.version numbers. Checks determine completeness of the downloaded data.
}
