% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_datasets.R
\name{ncbi_datasets_summary_genome}
\alias{ncbi_datasets_summary_genome}
\title{Summary of genomes present on NCBI}
\usage{
ncbi_datasets_summary_genome(
  accession = NULL,
  out.acc = "datasets_summary.ids",
  out.tsv = "datasets_summary.tsv",
  remove.intermediates = TRUE
)
}
\arguments{
\item{accession}{Character vector or file name containing accession.version or assembly.version numbers}

\item{out.tsv}{Name of TSV file (temporary)}

\item{remove.intermediates}{Boolean, whether to remove TSV file}
}
\description{
Wrapper for NCBI datasets CLI based on dataset summary (virus) genome.
Downloads genome reports and returns formatted output.
}
