% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_datasets.R
\name{ncbi_genome_files}
\alias{ncbi_genome_files}
\title{List NCBI genome dataset files}
\usage{
ncbi_genome_files(path = NULL)
}
\arguments{
\item{path}{Path to genome dataset (e.g. .../ncbi_dataset/data/GCA_000000000.1/)}
}
\description{
List files of NCBI genomes by category (CDS, genome, gbff, gff, gtf, proteins, Other)
}
