% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_genomeCollection.R
\name{nrow,genomeCollection-method}
\alias{nrow,genomeCollection-method}
\title{Get the number of genomes in a genomeCollection}
\usage{
\S4method{nrow}{genomeCollection}(x)
}
\description{
Get the number of genomes in a genomeCollection
}
