% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_Reads.R
\name{paired,Reads-method}
\alias{paired,Reads-method}
\title{Accessors for the FASTQ files of a Reads object.}
\usage{
\S4method{paired}{Reads}(x)
}
\description{
Accessors for the FASTQ files of a Reads object.
}
\author{
Oliver Dietrich
}
