% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_genomeCollection.R
\name{path,genomeCollection-method}
\alias{path,genomeCollection-method}
\title{Accessors for the 'path' element of an genomeCollection object.}
\usage{
\S4method{path}{genomeCollection}(x)
}
\description{
The \code{path} slot in an genomeCollection object holds
a character vector containing file paths.
}
\examples{
index(object)
}
\author{
Oliver Dietrich
}
