% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{phage_rbp_detect}
\alias{phage_rbp_detect}
\title{PhageRBPdetect}
\usage{
phage_rbp_detect(genome = NULL)
}
\arguments{
\item{genome}{FASTA}
}
\description{
Detect receptor binding proteins (RBP) in phage genomes
}
