% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{pharokka}
\alias{pharokka}
\title{Pharokka}
\usage{
pharokka(genome.fasta = NULL)
}
\arguments{
\item{genome.fasta}{Path to genome FASTA}
}
\description{
CLI wrapper for phage genome annotation using Pharokka.
}
