% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{plot_mummer_alignment}
\alias{plot_mummer_alignment}
\title{Plot MUMmer coordinates}
\usage{
plot_mummer_alignment(x, max.contigs = 10)
}
\arguments{
\item{x}{Data.frame with MUMmer coordinates}
}
\description{
Plot MUMmer coordinates
}
