% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{plot_survival}
\alias{plot_survival}
\title{Kaplan-Meier Survival Plots}
\usage{
plot_survival(object, time, censor, show_clusters = NULL)
}
\arguments{
\item{object}{A \code{GAUDIObject} containing cluster data.}

\item{time}{A named vector or data frame with a column named "time".}

\item{censor}{A named vector or data frame with a column named "censor".}

\item{show_clusters}{A numeric list of which clusters to plot, by default
plots all the clusters (except noise, cluster 0).}
}
\value{
A Kaplan-Meier plot, of ggsurvplot type.
}
\description{
This function creates a Kaplan-Meier survival plot and allows the user to
select which clusters to plot.
}
