% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_quality.R
\name{PlotFastqQuality}
\alias{PlotFastqQuality}
\title{PlotFastqQuality}
\usage{
PlotFastqQuality(object, read.col = "R1", rows = NULL)
}
\arguments{
\item{object}{genomeCollection}

\item{read.col}{Column in metadata(object) containing FASTQ paths}
}
\description{
View FASTQ quality
}
