% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads.R
\name{PlotReadsCounts}
\alias{PlotReadsCounts}
\title{PlotReadsCounts}
\usage{
PlotReadsCounts(
  object,
  name,
  label = "index",
  base.size = 20,
  label.size = 5,
  grid.cols = 1,
  top.point.size = 4,
  top.point.stroke = 1,
  jitter = FALSE,
  bottom.point.size = 3,
  bottom.point.stroke = 1,
  bottom.text.size = 10
)
}
\arguments{
\item{object}{genomeCollection}

\item{name}{Slot name of Reads object in genomeCollection}
}
\description{
Visualize FASTQ counts
}
