% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads.R
\name{PlotReadsQuality}
\alias{PlotReadsQuality}
\title{PlotReadsQuality}
\usage{
PlotReadsQuality(
  object,
  name,
  x = "length",
  y = "avg_quality",
  col = "type",
  wrap = "index",
  L.lenght_required = 100L,
  L.average_quality = 20L,
  L.percent_qualified = 50L,
  S.lenght_required = 30L,
  S.average_quality = 20L,
  S.percent_qualified = 50L,
  base.size = 20,
  pt.size = 0.5,
  pt.shape = 21,
  pt.stroke = 0.5,
  wrap.rows = NULL,
  wrap.cols = NULL
)
}
\arguments{
\item{object}{genomeCollection}

\item{read.col}{Column in metadata(object) containing FASTQ paths}
}
\description{
View FASTQ quality
}
