% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly.R
\name{polypolish}
\alias{polypolish}
\title{Polypolish}
\usage{
polypolish(
  assembly,
  out.dir,
  short.1 = NULL,
  short.2 = NULL,
  unpaired = NULL,
  threads = n_proc(),
  overwrite = FALSE
)
}
\arguments{
\item{assembly}{FASTA input assembly}

\item{out.dir}{Output folder}

\item{short.1}{FASTQ file of first short reads in each pair (R1)}

\item{short.2}{FASTQ file of second short reads in each pair (R2)}

\item{unpaired}{FASTQ file of unpaired short reads (S)}

\item{long}{FASTQ or FASTA file of long reads}
}
\description{
CLI wrapper for Polypolish for short-read polishing of long-read assemblies.
}
