% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{read_domtbl}
\alias{read_domtbl}
\title{Read HMMer domain table}
\usage{
read_domtbl(file)
}
\arguments{
\item{file}{Path to HMMer domain table}
}
\value{
Data.frame
}
\description{
Read HMMer domain table
}
