% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{read_gff3}
\alias{read_gff3}
\title{Read GFF3 file}
\usage{
read_gff3(
  file,
  extract.attributes = TRUE,
  keep.sequences = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{file}{Path to file}

\item{verbose}{Boolean, whether to display file header and object structure}

\item{keep_fasta_sequence_as_attributes}{Boolean value to indicate whether to keep
appended FASTA sequences as attributes to the returned data.frame}
}
\value{
data.frame
}
\description{
Read file formatted as GFF3 (https://gmod.org/wiki/GFF3#GFF3_Annotation_Section)
}
