% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Reads-class}
\alias{Reads-class}
\title{The Reads class}
\description{
The Read object stores file paths for raw and filtered FASTQ reads. Reads
can be either paired-end short reads (R1 + R2), single short reads (S) or
long reads (L). While samples can have multiple FASTQ files (e.g. hybrid
assemblies with L and R1+R2) each row corresponds to a sample. For multiple
sequencing runs of the same type (2x long reads) the FASTQ files must be
concatenated before being assigned to the Reads object or registered in
two distinct Reads object before being merged into a third Reads object.
}
\section{Slots}{

\describe{
\item{\code{R1}}{Character, file paths with forward paired-end reads}

\item{\code{R2}}{Character, file paths with reverse paired-end reads}

\item{\code{S}}{Character, file paths with unpaired reads}

\item{\code{L}}{Character, file paths with long reads}
}}

