% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_genomeCollection.R
\name{Reads,genomeCollection-method}
\alias{Reads,genomeCollection-method}
\title{Accessors to Reads objects contained in a genomeCollection object.}
\usage{
\S4method{Reads}{genomeCollection}(x, name = NULL)
}
\description{
Accessors to Reads objects contained in a genomeCollection object.
}
