% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads.R
\name{ReadsCounts}
\alias{ReadsCounts}
\title{ReadsCounts}
\usage{
ReadsCounts(
  object,
  name,
  recompute = FALSE,
  recompute.sample = NULL,
  multicore = TRUE,
  threads = n_proc(),
  debug = FALSE,
  debug.n = threads
)
}
\arguments{
\item{object}{genomeCollection}

\item{name}{Slot name of Reads object in genomeCollection}

\item{recompute}{Whether to re-compute the summary}

\item{multicore}{Whether to use mclapply for parallel excution}

\item{threads}{Integer, number of threads to use}

\item{debug}{Whether to run debugging mode: run as many samples as threads}
}
\description{
Count FASTQ files for a genomeCollection
}
