% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads.R
\name{ReadsQuality}
\alias{ReadsQuality}
\title{ReadsQuality}
\usage{
ReadsQuality(
  object,
  name,
  max.reads = 5000,
  min.base.quality = 15,
  multicore = TRUE,
  recompute = FALSE,
  recompute.sample = NULL,
  threads = n_proc(),
  debug = FALSE,
  debug.n = threads
)
}
\arguments{
\item{object}{genomeCollection}

\item{threads}{Integer, number of threads to use}

\item{debug}{Whether to run debugging mode: run as many samples as threads}

\item{read.col}{Column in metadata(object) containing FASTQ file paths}
}
\description{
Summarize quality of FASTQ files in a genomeCollection
}
