% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{RunBakta}
\alias{RunBakta}
\title{Run Bakta for a genomeCollection}
\usage{
RunBakta(
  object,
  regions = "regions",
  db = "../databases/bakta/",
  recompute = FALSE,
  recompute.sample = NULL,
  debug = FALSE,
  debug.n = 1
)
}
\arguments{
\item{object}{A genomeCollection object}

\item{recompute}{Whether to re-run Bakta even though an output GFF exists}

\item{debug}{Whether to run in debug mode (run only for debug.n samples)}

\item{debug.n}{Number of samples to run while debugging}
}
\description{
Run Bakta for a genomeCollection
}
