% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_quality.R
\name{RunFastp}
\alias{RunFastp}
\title{RunFastp}
\usage{
RunFastp(object, read.1, read.2, ...)
}
\arguments{
\item{object}{genomeCollection}

\item{read.1}{Name of column in metadata(object)}

\item{read.2}{Name of column in metadata(object)}

\item{...}{}
}
\description{
Preprocess FASTQ reads using fastp
}
