% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly.R
\name{RunFlye}
\alias{RunFlye}
\title{Run Flye}
\usage{
RunFlye(
  object,
  name = "Flye",
  reads.name = "filtered",
  type,
  recompute = FALSE,
  recompute.sample = NULL,
  samples.skip = NULL,
  debug = FALSE,
  debug.n = 1,
  ...
)
}
\arguments{
\item{object}{A genomeCollection object}

\item{name}{Slot to use for Assemblies(object)}

\item{reads.name}{Slot in ReadsNames(object}

\item{type}{Type of reads used for assembly. One of pacbio.raw, pacbio.corr, pacbio.hifi, nano.raw, nano.corr, nano.hq.}

\item{...}{Other arguments to flye_assembly()}

\item{threads}{Number of cores to use}
}
\description{
Run Flye assembly for a genomeCollection
}
