% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{RunPadloc}
\alias{RunPadloc}
\title{Run Padloc for a genomeCollection}
\usage{
RunPadloc(
  object,
  slot = "Bakta",
  db = "../databases/padloc",
  recompute = FALSE,
  recompute.sample = NULL,
  debug = FALSE,
  debug.n = 1
)
}
\arguments{
\item{object}{A genomeCollection object}

\item{slot}{A genomeObject slot storing genomes (FNA) OR gene (GFF) and protein (FAA) files}

\item{recompute}{Whether to re-run Padloc even though an output GFF exists}

\item{debug}{Whether to run in debug mode (run only for debug.n samples)}

\item{debug.n}{Number of samples to run while debugging}
}
\description{
Run Padloc for a genomeCollection
}
