% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{RunProdigal}
\alias{RunProdigal}
\title{Run Prodigal for a genomeCollection}
\usage{
RunProdigal(object, recompute = FALSE, debug = FALSE, debug.n = 1)
}
\arguments{
\item{object}{A genomeCollection object}

\item{recompute}{Whether to re-run Prodigal even though an output GFF exists}

\item{debug}{Whether to run in debug mode (run only for debug.n samples)}

\item{debug.n}{Number of samples to run while debugging}
}
\description{
Run Prodigal for a genomeCollection
}
