% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly.R
\name{RunSPAdes}
\alias{RunSPAdes}
\title{RunSPAdes}
\usage{
RunSPAdes(
  object,
  name = "SPAdes",
  type = "paired",
  reads.name = "filtered",
  recompute = FALSE,
  recompute.sample = NULL,
  ...
)
}
\arguments{
\item{object}{A genomeCollection object}

\item{name}{Slot to use for Assemblies(object)}

\item{type}{Type of reads to use for assembly. One of paired, unpaired.}

\item{reads.name}{Slot in ReadsNames(object}

\item{threads}{Number of cores to use}
}
\description{
Run SPAdes assembly for a genomeCollection
}
