% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly.R
\name{RunUnicycler}
\alias{RunUnicycler}
\title{Run Unicycler}
\usage{
RunUnicycler(
  object,
  name = "Unicycler",
  reads.name = "filtered",
  recompute = FALSE,
  recompute.sample = NULL,
  ...
)
}
\arguments{
\item{object}{A genomeCollection object}

\item{name}{Slot to use for Assemblies(object)}

\item{reads.name}{Slot in ReadsNames(object}

\item{type}{Type of reads to use for assembly. One of paired, unpaired.}

\item{threads}{Number of cores to use}
}
\description{
Run Unicycler assembly for a genomeCollection
}
