% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{skani_sketch}
\alias{skani_sketch}
\title{Skani sketch}
\usage{
skani_sketch(
  input,
  out.dir,
  individual.seqs = FALSE,
  threads = n_proc(),
  recompute = FALSE
)
}
\arguments{
\item{input}{Path(s) to FASTA files}

\item{out.dir}{Path to output directory}

\item{individual.seqs}{Whether to create sketches for individual sequences instead of entire files}

\item{threads}{Number of threads}
}
\description{
Sketch (index) genomes for fast, robust ANI calculation
}
