% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly.R
\name{skesa_assembly}
\alias{skesa_assembly}
\title{SKESA assembly}
\usage{
skesa_assembly()
}
\arguments{
\item{overwrite}{Boolean, whether to overwrite output}
}
\description{
CLI wrapper for SKESA assembly from short reads
}
