% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_genomeCollection.R
\name{[[,genomeCollection,ANY,missing-method}
\alias{[[,genomeCollection,ANY,missing-method}
\title{Subobjects and genome meta data}
\usage{
\S4method{[[}{genomeCollection,ANY,missing}(x, i, j, ...)
}
\arguments{
\item{x}{A genomeCollection object}

\item{i}{Name of genome meta data}

\item{j}{}
}
\description{
The \code{[[} operator pulls either subobjects
(e.g. assembly, annotation, distances, ...) or
genome meta data from a genomeCollection object
}
