% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_quality.R
\name{SummarizeFastqQuality}
\alias{SummarizeFastqQuality}
\title{SummarizeFastqQuality}
\usage{
SummarizeFastqQuality(object, read.col = "R1", threads = n_proc())
}
\arguments{
\item{object}{genomeCollection}

\item{read.col}{Column in metadata(object) containing FASTQ file paths}

\item{threads}{Integer, number of threads to use}
}
\description{
Summarize quality of FASTQ files in a genomeCollection
}
