% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classes.R
\docType{class}
\name{UBMIObject-class}
\alias{UBMIObject-class}
\title{UBMIObject Class}
\description{
A class representing the output of the UMAP-based Multi-omics Integration (UBMI) process.
The \code{UBMIObject} class encapsulates various results and metadata from the UBMI analysis,
including integrated factors, clustering results, silhouette scores, and metagene associations.
}
\details{
The \code{UBMIObject} class is designed to store comprehensive results from the UBMI analysis, providing an integrated view of multi-omics data.
It allows for easy access to both combined and individual omics data results, facilitating further analysis and visualization.
}
\section{Slots}{

\describe{
\item{\code{factors}}{A \code{data.frame} containing the UMAP factors for the integrated multi-omics data.}

\item{\code{clusters}}{A numeric vector representing the cluster assignments for each sample.}

\item{\code{silhouette_score}}{A numeric value representing the average silhouette score of the clustering.}

\item{\code{single_factors}}{A list of UMAP factors for individual omics datasets.}

\item{\code{metagenes_factor1}}{A \code{data.frame} containing the metagenes associated with the first dimension of the UMAP manifold.}

\item{\code{metagenes_factor1_rank}}{A character vector indicating the ranking of metagenes associated with the first dimension.}

\item{\code{metagenes_factor2}}{A \code{data.frame} containing the metagenes associated with the second dimension of the UMAP manifold.}

\item{\code{metagenes_factor2_rank}}{A character vector indicating the ranking of metagenes associated with the second dimension.}

\item{\code{single_metagenes_factor1}}{A list containing metagenes for individual omics datasets associated with the first dimension.}

\item{\code{single_metagenes_factor2}}{A list containing metagenes for individual omics datasets associated with the second dimension.}

\item{\code{ubmiVersion}}{A character string indicating the version of the UBMI package used for the analysis.}
}}

