% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{virusCollection-class}
\alias{virusCollection-class}
\title{The virusCollection Class}
\description{
The virusCollection object contains viral sequence data. Formatted based on NCBI datasets.
}
\section{Slots}{

\describe{
\item{\code{genomes}}{Data.frame containing genome metadata}

\item{\code{features}}{Data.frame containing genomic features (based on gff3)}

\item{\code{key}}{Character vector indicating the column to use as key between layout and data}

\item{\code{misc}}{A list of miscellaneous information}
}}

