% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly.R
\name{WriteAssemblyToGenome}
\alias{WriteAssemblyToGenome}
\title{Assembly summary}
\usage{
WriteAssemblyToGenome(
  object,
  name,
  filter = NULL,
  samples = NULL,
  min.size = 0,
  min.coverage = 0,
  max.size = Inf,
  return.filter = FALSE,
  scaffold = FALSE,
  scaffold.length = 50,
  debug = FALSE
)
}
\arguments{
\item{object}{genomeCollection}

\item{name}{Slot name of Assembly object in genomeCollection}

\item{filter}{Data.frame with filters to apply. Columns: sample, length.min, length.max, cov.min. (Not checked, experimental...)}

\item{samples}{Names of samples to write genomes for. Used to decide for different assemblies in different samples.}

\item{min.size}{Minimum size of contigs, either single value or threshold for each sample}

\item{min.coverage}{Minimum coverage of contigs, either single value or threshold for each sample}

\item{max.size}{Maximum size of contigs, either single value or threshold for each sample}

\item{return.filter}{Whether to return the filter data.frame. Used to adjust thresholds for individual samples and pass to 'filter' argument in second iteration.}

\item{scaffold}{Whether to create scaffolds from groups of contigs
param debug Whether to run in debug mode. Returns a list containing the contigs, filter, and sequences.}
}
\description{
Create a summary of the contigs contained in an Assembly object.
}
