% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{bakta}
\alias{bakta}
\title{Bakta annotation}
\usage{
bakta(
  input.fasta,
  output.dir,
  db = "../databases/bakta",
  prefix = "bakta",
  keep.contig.headers = TRUE,
  regions = NULL,
  download.db = FALSE,
  threads = n_proc(),
  force = FALSE
)
}
\arguments{
\item{input.fasta}{Path to input FASTA file}

\item{output.dir}{Path to output directory}

\item{db}{Character, path to the bakta database}

\item{regions}{Path to GFF3 file with pre-computed CDS regions}

\item{download.db}{Boolean, whether to download the database if missing}

\item{threads}{Integer, number of threads to use (default: all)}

\item{force}{Boolean, whether to force overwriting existing output folder}

\item{output.prefix}{Prefix of file names}
}
\description{
CLI wrapper for genome annotation using Bakta
}
