% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{CombineAnnotations}
\alias{CombineAnnotations}
\title{Combine annotations for a genomeCollection}
\usage{
CombineAnnotations(object, name, annotations, overwrite = FALSE)
}
\arguments{
\item{object}{A genomeCollection object}

\item{name}{Name of the slot in metadata(object)}

\item{annotations}{A character vector of annotations present in Annotations(object)}

\item{overwrite}{Whether to overwrite existing output files}
}
\description{
Combine the genes from multiple Annotation objects in a genomeCollection.
}
