% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{defense_finder}
\alias{defense_finder}
\title{DefenseFinder}
\usage{
defense_finder(
  file = NULL,
  out.dir = NULL,
  db.path = NULL,
  anti.defense = TRUE,
  anti.defense.only = FALSE,
  threads = n_proc(),
  update.db = FALSE
)
}
\arguments{
\item{file}{Character, path to the input file.}

\item{out.dir}{Character, path to target directory.}

\item{db.path}{Character, path to defense-system-model database (mentioned in
https://github.com/mdmparis/defense-finder/issues/8).}

\item{threads}{Numeric, number of threads.}

\item{update.db}{Boolean, whether to update the DefenseFinder database}
}
\description{
CLI wrapper for defense gene annotation using DefenseFinder
}
