% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{fasta_headers}
\alias{fasta_headers}
\title{Extract FASTA headers}
\usage{
fasta_headers(input.fasta = NULL)
}
\arguments{
\item{input.fasta}{Character, input FASTA file}
}
\description{
Extract FASTA headers
}
