% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_quality.R
\name{fastplong}
\alias{fastplong}
\title{Fastplong filtering of long reads}
\usage{
fastplong(
  file_in = NULL,
  file_out = NULL,
  length_required = 100L,
  mean_qual = 20L,
  unqualified_percent_limit = 40L,
  qualified_quality_phred = 15L,
  threads = n_proc(),
  overwrite = FALSE
)
}
\arguments{
\item{file_in}{Input file}

\item{file_out}{Output file}

\item{length_required}{Integer, reads shorter than length_required will be discarded, default is 15.}

\item{mean_qual}{Integer, if one read's mean_qual quality score < mean_qual, then this read is discarded. Default 0 means no requirement.}

\item{unqualified_percent_limit}{Integer, how many percents of bases are allowed to be unqualified (0~100). Default 40 means 40\%.}

\item{qualified_quality_phred}{Integer, the quality value that a base is qualified. Default 15 means phred quality >=Q15 is qualified.}

\item{threads}{Numeric, worker thread number, default is all (see n_proc for details)}

\item{overwrite}{}
}
\description{
CLI wrapper for fastplong to filter and trim long reads
}
