% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly.R
\name{flye_assembly}
\alias{flye_assembly}
\title{Flye assembly}
\usage{
flye_assembly(
  pacbio.raw = NA,
  pacbio.corr = NA,
  pacbio.hifi = NA,
  nano.raw = NA,
  nano.corr = NA,
  nano.hq = NA,
  genome.size = NULL,
  out.dir = NULL,
  log.file = NULL,
  threads = n_proc(),
  overwrite = FALSE
)
}
\arguments{
\item{pacbio.raw}{Path, PacBio regular CLR reads (<20\% error)}

\item{pacbio.corr}{Path, PacBio reads that were corrected with other methods (<3\% error)}

\item{pacbio.hifi}{Path, PacBio HiFi reads (<1\% error)}

\item{nano.raw}{Path, ONT regular reads, pre-Guppy5 (<20\% error)}

\item{nano.corr}{Path, ONT reads that were corrected with other methods (<3\% error)}

\item{nano.hq}{Path, ONT high-quality reads: Guppy5+ SUP or Q20 (<5\% error)}

\item{genome.size}{Character, estimated genome size (for example, 5m or 2.6g)}

\item{out.dir}{Path, Output directory}

\item{threads}{Integer,  number of parallel threads}
}
\description{
CLI wrapper for Flye to assemble genomes from long reads
}
