% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{geNomad}
\alias{geNomad}
\title{Detect MGEs with geNomad}
\usage{
geNomad(input.fasta, output.dir, db, force = FALSE)
}
\arguments{
\item{input.fasta}{Character, path to input FASTA}

\item{output.dir}{Character, path to output directory}

\item{db}{Character, path to database}

\item{force}{Boolean, whether to overwrite output.dir}
}
\description{
CLI wrapper for geNomad to detect mobile genetic elements (MGE)
}
