% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{get_pairwise_survival_data}
\alias{get_pairwise_survival_data}
\title{Kaplan-Meier Survival Summary Table}
\usage{
get_pairwise_survival_data(object, time, censor)
}
\arguments{
\item{object}{A \code{GAUDIObject} containing cluster data.}

\item{time}{A named vector or data frame with a column named "time".}

\item{censor}{A named vector or data frame with a column named "censor".

Centroids are calculated using median, quality is formed by a 80-20 split of
log p-value and scaled distance}
}
\value{
A data frame, with p-value of the survival comparison, distance
between clusters in the GAUDI manifold, and quality, a non-interpretable
metric to assist in prioritizing clusters to target.
}
\description{
This function provides information on pairwise comparisons between clusters
of the GAUDI object.
}
\examples{
table <- get_pairwise_survival_data(object, time, censored)
plot_survival(object, time, censored, table[["clusters"]][[1]])

}
