% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_mer_utils.R
\name{jellyfish_count_kmers}
\alias{jellyfish_count_kmers}
\title{Count nucleotide k-mers using jellyfish}
\usage{
jellyfish_count_kmers(
  infile = NULL,
  sequence = NULL,
  outfile = NULL,
  k = 31,
  threads = NULL,
  return.counts = TRUE
)
}
\arguments{
\item{k}{Size of k-mer}

\item{threads}{Number of cores to use (default: all)}

\item{seq}{Nucleotide sequence(s) to count k-mers. Input must be character vector or list.}
}
\description{
Count nucleotide k-mers using jellyfish
}
