% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_mer_utils.R
\name{jellyfish_count_kmers}
\alias{jellyfish_count_kmers}
\title{Jellyfish k-mer counting}
\usage{
jellyfish_count_kmers(
  input.sequence = NULL,
  output.tsv = NULL,
  temp.dir = "tmp/jellyfish_count/",
  k = 32,
  hash.size = "100M",
  threads = n_proc(),
  remove.intermediates = TRUE
)
}
\arguments{
\item{input.sequence}{File path}

\item{output.tsv}{File path}

\item{k}{Integer, length of k-mer}

\item{hash.size}{Character, Initial hash size (e.g. '100M')}

\item{threads}{Integer, Number of threads}
}
\description{
CLI wrapper for Jellyfish to count k-mers
}
