% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{mash_distance}
\alias{mash_distance}
\title{Compute Mash distance for genomic sequences}
\usage{
mash_distance(
  query,
  reference = NULL,
  out.dir,
  k = 21,
  s = 10000,
  alphabet = "nucleotide",
  threads = n_proc(),
  recompute = FALSE
)
}
\arguments{
\item{query}{XStringSet or vector of file paths.}

\item{reference}{XStringSet or vector of file paths.}

\item{k}{Integer, k-mer size}

\item{s}{Integer, sketch size}

\item{alphabet}{Character, use 'nucleotide' (ACGT) or 'aminoacid' (A-Z, except BJOUXZ) alphabet}

\item{threads}{Integer, this many threads will be spawned for processing}

\item{output.dir}{Path to output directory}

\item{individual.seqs}{Boolean, whether to sketch individual sequences rather than whole files}

\item{recompute.sketch}{Boolean, whether to re-compute the sketch file}

\item{recompute.dist}{Boolean, whether to re-compute the Mash distances}
}
\description{
Estimate sequence identity (average nucleotide identity, ANI) and containment
(alignment fraction, AF) for biological sequences. Input can be either Biostrings
objects (DNAStringSet, AAStringSet) or character vector of FASTA file paths. If
no reference sequences are supplied, query sequences will be compared to each other.
Sequence names will be derived from ...
}
