% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{mash_distance}
\alias{mash_distance}
\title{CLI wrapper for Mash}
\usage{
mash_distance(
  input = NULL,
  output.dir = NULL,
  k = 21,
  s = 1000,
  individual.seqs = FALSE,
  threads = n_proc(),
  recompute.sketch = FALSE,
  recompute.dist = FALSE
)
}
\arguments{
\item{input}{Character, vector of input FASTA sequences. If names are present, will be used for output IDs.}

\item{k}{Integer, k-mer size}

\item{s}{Integer, sketch size}

\item{threads}{Integer, this many threads will be spawned for processing.}

\item{recompute.sketch}{Boolean, whether to re-compute the sketch file}

\item{recompute.dist}{Boolean, whether to re-compute the Mash distances}

\item{output}{Path, output prefix (first input file used if unspecified). The suffix '.msh' will be appended.}

\item{individual}{Boolean, whether to sketch individual sequences rather than whole files}

\item{sketch.archive}{File name, destination of the sketch archive}
}
\description{
Compute Mash distance for genomic sequences
}
