% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{mash_sketch_distance}
\alias{mash_sketch_distance}
\title{Compute Mash distance between pre-computed sketches}
\usage{
mash_sketch_distance(
  query,
  reference = NULL,
  out.file,
  symmetric = FALSE,
  threads = n_proc()
)
}
\arguments{
\item{query}{Character, path to query sketch}

\item{reference}{Character, path to reference sketch}

\item{out.file}{Character, file name to output.tsv}

\item{symmetric}{Whether to also compare reference to query, in contrast to only query to ref}
}
\description{
Compute Mash distance between pre-computed sketches
}
