% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_datasets.R
\name{ncbi_genome_summary}
\alias{ncbi_genome_summary}
\title{Summary of genomes present on NCBI}
\usage{
ncbi_genome_summary(
  dirname = "data/",
  refseq = "refseq_assembly_summary.txt",
  genbank = "genbank_assembly_summary.txt",
  url_refseq = "ftp://ftp.ncbi.nlm.nih.gov/genomes/refseq/bacteria/assembly_summary.txt",
  url_genbank = "ftp://ftp.ncbi.nlm.nih.gov/genomes/genbank/assembly_summary_genbank.txt",
  max.download.minutes = 30,
  max.download.seconds = max.download.minutes * 60
)
}
\arguments{
\item{dirname}{Name of directory to save assembly summaries (default: data/)}

\item{refseq}{Name of refseq assembly summary}

\item{genbank}{Name of genbank assembly summary}

\item{url_refseq}{URL to NCBI FTP for refseq summary}

\item{url_genbank}{URL to NCBI FTP for genbank summary}

\item{max.download.minutes}{Numeric, maximum download time in seconds}
}
\description{
Download assembly summary for GenBank and RefSeq from NCBI FTP.
}
