% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{padloc}
\alias{padloc}
\title{PADLOC: Prokaryotic Antiviral Defence LOCator}
\usage{
padloc(
  protein.faa = NULL,
  genes.gff = NULL,
  genome.fna = NULL,
  crispr.gff = NULL,
  out.dir,
  db.path,
  fix.prodigal = FALSE,
  overwrite = FALSE,
  threads = n_proc()
)
}
\arguments{
\item{protein.faa}{Character, path to protein FASTA (only in combination with genes.gff)}

\item{genes.gff}{Character, path to genomic GFF (only in combination with protein.faa)}

\item{genome.fna}{Character, path to genomic FASTA}

\item{crispr.gff}{Character, path to CRISPRfinder GFF.}

\item{db.path}{Character, path to padloc database.}

\item{overwrite}{Boolean, whether to overwrite existing results.}

\item{threads}{Numeric, number of threads. Defaults to nproc.}
}
\description{
CLI wrapper for defense gene annotation using PADLOC. Uses either a genome FASTA or
a combination of protein FASTA with genes GFF. Database will be downloaded if not present.
}
