% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_Reads.R
\name{paired,Reads-method}
\alias{paired,Reads-method}
\title{Accessors for the FASTQ files of a Reads object.}
\usage{
\S4method{paired}{Reads}(x)
}
\description{
The \code{meta.data} slot in an genomeCollection object holds
a data.frame containing unstructured information associated to each genome.
}
\author{
Oliver Dietrich
}
