% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly.R
\name{raven_assembly}
\alias{raven_assembly}
\title{Raven assembly}
\usage{
raven_assembly(input.fastq, output.fasta, threads = n_proc())
}
\arguments{
\item{input.fastq}{Fastq files to assemble}

\item{output.fasta}{Fasta file for assembly}

\item{threads}{Number of CPU cores to use}
}
\description{
CLI wrapper for Raven to assemble genomes from long reads
}
