% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{read_assembly_graph}
\alias{read_assembly_graph}
\title{Read assembly graph}
\usage{
read_assembly_graph(input.graph = NULL, remove.sequences = TRUE)
}
\arguments{
\item{input.graph}{Character, path to input file}
}
\description{
Read assembly graph from .gfa file. File specification is available at https://github.com/GFA-spec/GFA-spec.
}
