% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_Reads.R
\name{AddReads}
\alias{AddReads}
\title{Add a \code{Reads} object to a \code{\link{genomeCollection}} object}
\usage{
AddReads(
  object = NULL,
  name = NULL,
  x = NULL,
  ids = "ID",
  R1 = "R1",
  R2 = "R2",
  S = "S",
  L = "L"
)
}
\arguments{
\item{object}{A \code{\link{genomeCollection}} object}

\item{x}{Data.frame containing the paths to reads files}

\item{R1}{Name of the column containing file paths with forward paired-end reads}

\item{R2}{Name of the column containing file paths with reverse paired-end reads}

\item{S}{Name of the column containing file paths with unpaired reads}

\item{L}{Name of the column containing file paths with long reads}
}
\value{
A \code{\link{genomeCollection}} object
}
\description{
Create a \code{Reads} object from a data.frame with defined components
and add it to the reads slot of a genomeCollection object.
}
