% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{RunGenomad}
\alias{RunGenomad}
\title{Run GeNomad for a genomeCollection}
\usage{
RunGenomad(
  object,
  recompute = FALSE,
  recompute.sample = NULL,
  debug = FALSE,
  debug.n = 1,
  ...
)
}
\arguments{
\item{object}{A genomeCollection object}
}
\description{
Run GeNomad for a genomeCollection
}
