% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly.R
\name{skesa_assembly}
\alias{skesa_assembly}
\title{SKESA assembly}
\usage{
skesa_assembly()
}
\arguments{
\item{input.1}{Character, file name of forward paired-end reads}

\item{input.2}{Character, file name of reverse paired-end reads}

\item{input.s}{Character, file name of unpaired reads}

\item{threads}{Integer, number of threads to use}

\item{overwrite}{Boolean, whether to overwrite output}
}
\description{
CLI wrapper for SKESA assembly from short reads
}
