% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly.R
\name{spades_assembly}
\alias{spades_assembly}
\title{SPAdes assembly}
\usage{
spades_assembly()
}
\arguments{
\item{overwrite}{Boolean, whether to overwrite output}
}
\description{
CLI wrapper for SPAdes assembly from short reads
}
