% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly.R
\name{spades_assembly}
\alias{spades_assembly}
\title{SPAdes assembly}
\usage{
spades_assembly(
  input.1 = NULL,
  input.2 = NULL,
  input.s = NULL,
  out.dir = NULL,
  isolate = FALSE,
  single.cell = FALSE,
  metagenomic = FALSE,
  biosynthetic = FALSE,
  sewage = FALSE,
  corona = FALSE,
  rna = FALSE,
  plasmid = FALSE,
  metaviral = FALSE,
  metaplasmid = FALSE,
  rnaviral = FALSE,
  log.file = NULL,
  threads = n_proc(),
  overwrite = FALSE
)
}
\arguments{
\item{input.1}{File name, file with forward paired-end reads}

\item{input.2}{File name, file with reverse paired-end reads}

\item{input.s}{File name, file with unpaired reads}

\item{out.dir}{Path to output directory}

\item{threads}{Integer, number of threads to use}

\item{overwrite}{Boolean, whether to overwrite output}
}
\description{
CLI wrapper for SPAdes assembly from short reads
}
