% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_quality.R
\name{summarize_fastq_quality}
\alias{summarize_fastq_quality}
\title{Compute FASTQ summary}
\usage{
summarize_fastq_quality(
  in_file = NULL,
  min.base.quality = 15,
  max.reads = 5000
)
}
\arguments{
\item{in_file}{File name.}

\item{min.base.quality}{PHRED score of individual bases to be qualified (impacts percent qualified).}

\item{max.reads}{Integer, number of reads to sample}
}
\description{
Compute FASTQ summary
}
