% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads.R
\name{summarize_fastq_quality}
\alias{summarize_fastq_quality}
\title{Compute FASTQ summary}
\usage{
summarize_fastq_quality(
  in_file = NULL,
  max.reads = 5000,
  min.base.quality = 15,
  reads.present = character()
)
}
\arguments{
\item{in_file}{File name.}

\item{max.reads}{Integer, number of reads to sample}

\item{min.base.quality}{PHRED score of individual bases to be qualified (impacts percent qualified).}
}
\description{
Compute average PHRED score across a subset of reads within a FASTQ file
}
