% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reads.R
\name{summarize_read_quality}
\alias{summarize_read_quality}
\title{Compute read quality}
\usage{
summarize_read_quality(
  fastq = NULL,
  read = NULL,
  min.base.quality = 15,
  per.base.quality = FALSE
)
}
\arguments{
\item{fastq}{Character, path to FASTQ file}

\item{read}{Integer, which read to summarize}

\item{min.base.quality}{Integer, PHRED score of individual bases to be qualified}

\item{per.base.quality}{Boolean, whether to return per-base quality scores (TRUE) or summary scores (FALSE)}
}
\description{
Compute average PHRED score for a single read within a FASTQ file
}
