% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qc.R
\name{summarize_read_quality}
\alias{summarize_read_quality}
\title{Compute read quality}
\usage{
summarize_read_quality(
  fastq = NULL,
  read = NULL,
  min.base.quality = 15,
  per.base.quality = FALSE
)
}
\description{
Compute read quality
}
