% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{system3}
\alias{system3}
\title{Invoke a System Command}
\usage{
system3(
  command,
  max.print = 25,
  log.file = NULL,
  verbose = TRUE,
  include.errors = TRUE
)
}
\arguments{
\item{command}{the system command to be invoked, as a character string.}
}
\description{
system3 invokes the OS command specified by command.
}
