% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubmi.R
\name{ubmi}
\alias{ubmi}
\title{UMAP-Based Multi-Omics Data Integration}
\usage{
ubmi(
  omics,
  umap_params = list(n_neighbors = 15, n_components = 4, pca = min(nrow(omics[[1]]),
    ncol(omics[[1]]))),
  umap_params_conc = list(n_neighbors = 15, n_components = 2),
  min_pts = 5,
  xgboost_params = list(lambda = 0, eta = 0.5, gamma = 50, max_depth = 10, subsample =
    0.95),
  compute_features = TRUE,
  combine_omics = FALSE,
  clean_feature_names = FALSE,
  samples_in_rows = TRUE
)
}
\arguments{
\item{omics}{A list of omics data matrices, with each matrix representing a different omics data type.}

\item{umap_params}{A list of parameters for individual UMAP embeddings. Defaults to \code{list(n_neighbors = 15, n_components = 4)}.}

\item{umap_params_conc}{A list of parameters for concatenated UMAP embeddings. Defaults to \code{list(n_neighbors = 15, n_components = 2)}.}

\item{min_pts}{The minimum number of points for a cluster in HDBSCAN. Defaults to 10.}

\item{xgboost_params}{A list of parameters for XGBoost modeling. Defaults to \code{list(lambda = 1, eta = 0.3, gamma = 100, max_depth = 10, subsample = 0.95)}.}

\item{compute_features}{Logical. Indicates whether to compute metagenes using XGBoost. Defaults to TRUE.}

\item{samples_in_rows}{Logical. Indicates whether samples are in rows (TRUE) or columns (FALSE) of the omics matrices. Defaults to TRUE.}
}
\value{
A \code{UBMIObject} containing the results of the integration, including factors, clusters, silhouette scores, and metagenes.
}
\description{
\code{ubmi} performs unsupervised integration of multi-omics data.
It utilizes Uniform Manifold Approximation and Projection (UMAP) for dimensionality reduction and HDBSCAN for clustering.
The method concatenates individual UMAP embeddings of each omics data type, followed by a second UMAP for integrated analysis.
Clustering is performed on the integrated data, and optionally, cluster 0 reassignment and feature computation can be done.
}
\examples{
# Assuming expr, meth, miRNA, and metab are your omics datasets
omics_data <- list(expr, meth, miRNA, metab)
ubmi_result <- ubmi(omics = omics_data)

}
