% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly.R
\name{unicycler_assembly}
\alias{unicycler_assembly}
\title{Unicycler assembly}
\usage{
unicycler_assembly(
  short.1 = NULL,
  short.2 = NULL,
  unpaired = NULL,
  long = NULL,
  out.dir = NULL,
  log.file = NULL,
  keep = 1,
  mode = "normal",
  min_fasta_length = 100,
  linear_seqs = 0,
  threads = n_proc(),
  overwrite = FALSE
)
}
\arguments{
\item{short.1}{FASTQ file of first short reads in each pair (R1)}

\item{short.2}{FASTQ file of second short reads in each pair (R2)}

\item{unpaired}{FASTQ file of unpaired short reads (S)}

\item{long}{FASTQ or FASTA file of long reads}

\item{out.dir}{Output directory}

\item{keep}{Level of file retention 0 = only keep final files: assembly (FASTA, GFA and log),
1 = also save graphs at main checkpoints, 2 = also keep SAM (enables fast rerun in different mode),3 = keep all temp files and save all graphs (for debugging)}

\item{mode}{Bridging mode: conservative, normal, bold (default: normal)}

\item{min_fasta_length}{Exclude contigs from the FASTA file which are shorter than this length (default: 100)}

\item{linear_seqs}{The expected number of linear (i.e. non-circular) sequences in the underlying sequence (default: 0)}

\item{threads}{Number of threads used}
}
\description{
CLI wrapper for Unicycler genome assembly.
Ideally hybrid (long + short) but accepts both short and long reads only.
}
