% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{write_gff3}
\alias{write_gff3}
\title{Write GFF3 file}
\usage{
write_gff3(x, filename, replace.attributes = FALSE)
}
\arguments{
\item{x}{Data.frame containing GFF3 formatted columns}

\item{filename}{File path}

\item{replace.attributes}{Whether to replace the attributes column
with tag-value pairs gather from non-GFF columns.}
}
\description{
Write data.frame to GFF3 formatted file
}
