% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_gff3}
\alias{write_gff3}
\title{Write GFF3 file}
\usage{
write_gff3()
}
\arguments{
\item{object}{Data.frame}

\item{file}{Character, file path}
}
\description{
Write data.frame to GFF3 formatted file
}
